+++ app/src/main/java/eu/kanade/presentation/track/TrackerSearchPreviewProvider.kt
@@ -62,7 +62,7 @@ internal class TrackerSearchPreviewProvider : PreviewParameterProvider<@Composab
     private fun randTrackSearch() = TrackSearch().let {
         it.id = Random.nextLong()
         it.manga_id = Random.nextLong()
-        it.tracker_id = Random.nextInt()
+        it.tracker_id = Random.nextLong()
         it.remote_id = Random.nextLong()
         it.library_id = Random.nextLong()
         it.title = lorem((1..10).random()).joinToString()
